Composer based packaging with psr-4 autoloader
