+++ .gitignore
@@ -1,0 +1,3 @@
+vendor/*
+.idea/*
+test.php
