+++ src/ShuttleException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ToCreateStudio\ShuttleExport;
+
+
+class ShuttleException extends \Exception
+{
+}
+
+;
